requirements.txt
celery==4.4.2
cloudinary==1.20.0
Django==2.2.12
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cloudinary-storage==0.2.3
django-heroku==0.3.1
django-mptt==0.11.0
gunicorn==20.0.4
Pillow==7.1.1
python-dotenv==0.13.0
requests==2.23.0
whitenoise==5.0.1
